requirements.txt
asgiref==3.2.7
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-dotenv==1.4.2
django-filter==2.0.0
django-url-filter==0.3.15
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
enum-compat==0.0.3
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.4
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
